Add specification for importance factor reporting
The commonly calculated importance factors (DIF, MIF, ...)
can be reported in attributes of variables (basic events).
The approach has been tested with SCRAM
and is flexible for new factors or omitted factors.

Note that importance factor reporting is missing for CCF events
because the general representation of them
is not yet accepted into the report layer.
It should be trivial to add CCF event importance factor reporting
(if not cumbersome, though)
after the initial formatting is established.