Add support for reporting histogram (density)
This modifies the "quantiles" schema to support bins
similar to the MEF input.
The distribution/frequency of the data is meant to be in the histogram
as a statistical measure for reporting.

It would be somewhat beneficial
if histogram reporting schema matched exactly
the input histogram-distribution (deviate) schema
since the semantics is the same.