Limit XOR and IFF to two arguments
Currently, XOR and IFF are specified as n-ary connectives.
In addition to being rarely used,
these connectives are not intuitive in their associative, n-ary forms.
For example, a naive user may confuse XOR with natural 'either .. or',
which doesn't hold with more than 2 args;
it is rather 'odd' (pun intended :)).

I think those FTA tools that do support these gates
limit the number of arguments to 2 (e.g., graphically).
It would be unreasonable to require generality without a practical need.